+++ src/ui/views/buttons.jsx
@@ -17,41 +17,33 @@ import {FontStyle} from '../elements/font-style.jsx';
 import {FontWeight} from '../elements/font-weight.jsx';
 import {FontColor} from '../elements/font-color.jsx';
 import {FontSize} from '../elements/font-size.jsx';
+import {BackgroundColor} from '../elements/bg-color.jsx';
 import {FontLineHeight} from '../elements/font-lineheight.jsx';
 
-export function FontView(){
-    return <Container className="ejs-customizer-fonts">
-        <Group title="Global">
-            <FontFamily selector={selectors.container} />
-            <FontStyle selector={selectors.container} />
-            <FontWeight selector={selectors.container} />
-            <FontColor selector={selectors.container} />
-            <FontSize selector={selectors.container} />
-            <FontLineHeight selector={selectors.container} />
-        </Group>
-
-        <Group title="Line numbers">
-            <FontFamily selector={selectors.linenumbers} />
-            <FontStyle selector={selectors.linenumbers} />
-            <FontWeight selector={selectors.linenumbers} />
-            <FontColor selector={selectors.linenumbers} />
-            <FontSize selector={selectors.linenumbers} />
-        </Group>
-
+export function ButtonView(){
+    return <Container className="ejs-customizer-buttons">
         <Group title="Generic buttons">
-            <FontFamily selector={selectors.buttons.generic}/>
+            <FontFamily selector={selectors.buttons.generic} />
+            <FontSize selector={selectors.buttons.generic} />
+            <FontLineHeight selector={selectors.buttons.generic} />
+
             <FontStyle selector={selectors.buttons.generic} />
             <FontWeight selector={selectors.buttons.generic} />
+           
             <FontColor selector={selectors.buttons.generic} />
-            <FontSize selector={selectors.buttons.generic} />
+            <BackgroundColor selector={selectors.buttons.generic} />
         </Group>
 
         <Group title="Codegroup buttons">
-            <FontFamily selector={selectors.buttons.codegroup}/>
+            <FontFamily selector={selectors.buttons.codegroup} />
+            <FontSize selector={selectors.buttons.codegroup} />
+            <FontLineHeight selector={selectors.buttons.codegroup} />
+
             <FontStyle selector={selectors.buttons.codegroup} />
             <FontWeight selector={selectors.buttons.codegroup} />
+           
             <FontColor selector={selectors.buttons.codegroup} />
-            <FontSize selector={selectors.buttons.codegroup} />
+            <BackgroundColor selector={selectors.buttons.codegroup} />
         </Group>
     </Container>;
 }