+++ src/ui/views/lines.jsx
@@ -12,10 +12,27 @@ import * as React from 'dom-magic';
 import {selectors} from '../../customizer/selectors';
 import {Container} from '../components/container.jsx';
 import {Group} from '../components/group.jsx';
+import {FontFamily} from '../elements/font-family.jsx';
+import {FontStyle} from '../elements/font-style.jsx';
+import {FontWeight} from '../elements/font-weight.jsx';
+import {FontColor} from '../elements/font-color.jsx';
+import {FontSize} from '../elements/font-size.jsx';
 import {BackgroundColor} from '../elements/bg-color.jsx';
 
 export function LineView(){
     return <Container className="ejs-customizer-lines">
+
+        <Group title="Line numbers">
+            <FontFamily selector={selectors.linenumbers} />
+            <FontSize selector={selectors.linenumbers} />
+
+            <FontStyle selector={selectors.linenumbers} />
+            <FontWeight selector={selectors.linenumbers} />
+           
+            <FontColor selector={selectors.linenumbers} />
+            <BackgroundColor selector={selectors.linenumbers} />
+        </Group>
+
         <Group title="Lines hover">
            <BackgroundColor selector={selectors.linehover} />
         </Group>